Various fixes

+++ PiranhaPlant.cpp
@@ -15,8 +15,8 @@
  *	@param y posizione lungo l'asse y
  *	@param textures puntatore al vettore delle texture da utilizzare
  */
-PiranhaPlant::PiranhaPlant(int x, int y, std::vector<sf::Texture> *textures)
-	: Enemy(PIRANHA_PLANT, x, y, SIZE, PIRANHAPLANT_HEIGHT, ((*textures)[0]))
+PiranhaPlant::PiranhaPlant(float x, float y, std::vector<sf::Texture> *textures) 
+: Enemy(PIRANHA_PLANT, x, y, SIZE, PIRANHAPLANT_HEIGHT, ((*textures)[0]))
 {
 	this -> textures = textures;
 	this -> textureID = PLANT_0;
